+++ SRC/Public/Extensions/MS_DI/IServiceContainerExtensions.cs
@@ -18,7 +18,7 @@ namespace Solti.Utils.DI
         /// <param name="provider">The newly created <see cref="IServiceProvider"/> instance.</param>
         /// <returns>The scope of the newly created provider.</returns>
         /// <exception cref="InvalidOperationException">There are one or more abstract entries in the collection.</exception>
-        public static IDisposable CreateProvider(this IServiceContainer self, out IServiceProvider provider)
+        public static Disposable CreateProvider(this IServiceContainer self, out IServiceProvider provider)
         {
             Ensure.Parameter.IsNotNull(self, nameof(self));