+++ history.md
@@ -108,4 +108,5 @@
   - *fixed:* named services could not be queried via `ServiceProvider` ([details](https://github.com/Sholtee/injector/commit/bd97a0caf2763b3a28d4c511b2ee02006b072b3b ))
 - 2.0.0:
   - *breaking:* In case of not supported interface a `NotSupportedException` is thrown (instead of `InvalidOperationException`)
+  - *breaking:* To support async scope disposal `IServiceContainerExtensions.CreateProvider()` returns `Disposable` (which implements `IDisposable` and `IAsyncDisposable`)
   - *improved:* `Injector` creation performance